+++ climpred/options.py
@@ -8,23 +8,29 @@ _VALIDATORS = {
 
 
 class set_options:
-    """Set options for climpred in a controlled context. Similar to xr.set_otions.
+    """Set options for climpred in a controlled context. Analogous to
+    `xarrayset_options(**option) <http://xarray.pydata.org/en/stable/generated/xarray.set_options.html>`_.
 
     Currently supported options:
-    - ``seasonality``: maximum display width for ``repr`` on xarray objects.
-      Default: ``dayofyear``.
 
-    You can use ``set_options`` either as a context manager:
-    >>> kw = dict(metric='mse', comparison='e2o', dim='init', alignment='same_verifs',
-    ...           reference='climatology')
-    >>> with climpred.set_options(seasonality='month'):
-    ...     HindcastEnsemble.verify(**kw).SST.sel(skill='climatology')
-    <xarray.DataArray 'SST' (lead: 10)>
-    array([0.03712573, 0.03712573, 0.03712573, 0.03712573, 0.03712573,
-           0.03712573, 0.03712573, 0.03712573, 0.03712573, 0.03712573])
-    Coordinates:
-      * lead     (lead) int32 1 2 3 4 5 6 7 8 9 10
-        skill    <U11 'climatology'
+        * ``seasonality``: Attribute to group dimension ``groupby(f"{dim}.{seasonality}"")``.
+            Used in ``reference=climatology`` and :py:meth:`~climpred.classes.HindcastEnsemble.remove_bias`.
+                - Allowed: ["dayofyear", "weekofyear", "month"]
+                - Default: ``dayofyear``.
+
+    Examples:
+        You can use ``set_options`` either as a context manager:
+
+        >>> kw = dict(metric='mse', comparison='e2o', dim='init', alignment='same_verifs',
+        ...           reference='climatology')
+        >>> with climpred.set_options(seasonality='month'):
+        ...     HindcastEnsemble.verify(**kw).SST.sel(skill='climatology')
+        <xarray.DataArray 'SST' (lead: 10)>
+        array([0.03712573, 0.03712573, 0.03712573, 0.03712573, 0.03712573,
+               0.03712573, 0.03712573, 0.03712573, 0.03712573, 0.03712573])
+        Coordinates:
+          * lead     (lead) int32 1 2 3 4 5 6 7 8 9 10
+            skill    <U11 'climatology'
     """
 
     def __init__(self, **kwargs):