+++ climpred/reference.py
@@ -135,6 +135,8 @@ def compute_climatology(
         clim (xarray object): Results of climatology forecast with the input metric
             applied.
     """
+    seasonality_str = OPTIONS["seasonality"]
+
     if isinstance(dim, str):
         dim = [dim]
     # Check that init is int, cftime, or datetime; convert ints or cftime to datetime.
@@ -161,14 +163,14 @@ def compute_climatology(
 
     if kind == "perfect":
         forecast, verif = comparison.function(hind, metric=metric)
-        climatology_day = verif.groupby("init.dayofyear").mean()
+        climatology_day = verif.groupby(f"init.{seasonality_str}").mean()
     else:
         forecast, verif = comparison.function(hind, verif, metric=metric)
-        climatology_day = verif.groupby("time.dayofyear").mean()
+        climatology_day = verif.groupby(f"time.{seasonality_str}").mean()
 
     climatology_day_forecast = climatology_day.sel(
-        dayofyear=forecast.init.dt.dayofyear, method="nearest"
-    ).drop("dayofyear")
+        {seasonality_str:getattr(forecast.init.dt, seasonality_str)}, method="nearest"
+    ).drop(seasonality_str)
 
     if kind == "hindcast":
         climatology_day_forecast = climatology_day_forecast.rename({"init": "time"})